Add files via upload
The start of the project was okay, but I ran into problems when I got held up with other school stuff. Even with extra time, I struggled with some tech issues like MySQL and XAMPP. Because of all that, I couldn't make the app like I imagined. If I did it again, I'd make a simpler diary app and not try to cram so much in, especially with other schoolwork going on.